Add dry-run option to parse template
We need to have a ability to validate the template before uploading the
AWS. This will enable faster feedback loop for CI build on pull request.

This commit adds a new dry-run option to the `ecs-upload-task` command.

Usage:
```
ecs-upload-task --file taskdefinition --dry-run
```

It will return either a successful message or a parsed error.